trainer/requirements.txt
tensorflow==2.19.0
keras==3.10.0
numpy==2.0.2
pillow==11.3.0
psycopg2-binary
scikit-learn